Fix a bug in cleanup code
Steve Deitz found a bug in the cleanup code after we started looking into the delete of an in-progress creation theory. This pass should be continue - this code will delete instances before they're RUNNING, leading to the create operation returning a 409 CONFLICT (because the operation was canceled by the delete).

This bug causing 409s will be easily caused by running two copies of this benchmark at the same time against the same project.