Fix potential causes of 409 errors (#5)
* Fix potential causes of 409 errors

Fixing two potential causes of 409 errors.
The GCE API returns a 409 error on instances.insert when the name is already in use.
Looking at this code this is either due to calling insert twice in the same second, or due to internal retries, either in the googleapis python code, or internal retries once Google has received the HTTP request.

This change adds a UUID into the instance name to make it unique, and uses that UUID as the requestId for the instances.insert call. This makes the call idempotent, and retries will all return the same operation rather than returning a 409.

* Fix a bug in cleanup code

Steve Deitz found a bug in the cleanup code after we started looking into the delete of an in-progress creation theory. This pass should be continue - this code will delete instances before they're RUNNING, leading to the create operation returning a 409 CONFLICT (because the operation was canceled by the delete).

This bug causing 409s will be easily caused by running two copies of this benchmark at the same time against the same project.